Avoid reuse of the variable 'view', fixing a bug in list update
An update which caused a new element to be inserted and existing elements to be
removed would cause the newly added element to be removed repeatedly while
traversing the to-be-removed traversed elements. For example:

function Item() { this.el = el('div'); }
var items = list('div', Item);
items.update([ 1, 2, 3 ]);
items.update([ 1 ]);

-> NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be
removed is not a child of this node.